Major refactor WIP
Cleaning up a lot of terrible architecture that was original thrown together for testing.

I am hoping this architecture will enable easy implementation of AABBs and probably BVHs in the future, as well as better performance in upcoming commits due to rendering cuboids as a singular primitive instead of a compounded shape made from 6 quads.

Cuboids are not yet implemented and a lot of the shader side stuff isn't done yet like traceable groups but that's to come