Handle internal rays in cuboids
Before, we were getting stuck in an infinite loop of reflecting/refracting when a ray entered a cuboid, causing immense performance problems as the ray would just go until max depth 100% of the time. It also didn't render properly, basically just becoming a black hole for rays.

This is finally fixed I think, admittedly this is a bit crude and a ray marching approach may be better long term but I'm just glad to get it working for now.

Small possibility this isn't perfect but it's surprisingly hard to tell if it's being modelled correctly